Manual merge from mysql-5.1 to mysql-5.5

+++ sql/sql_plugin.cc
@@ -714,7 +714,7 @@ bool plugin_is_ready(const LEX_STRING *name, int type)
 }
 
 
-SHOW_COMP_OPTION plugin_status(const char *name, int len, size_t type)
+SHOW_COMP_OPTION plugin_status(const char *name, size_t len, int type)
 {
   LEX_STRING plugin_name= { (char *) name, len };
   return plugin_status(&plugin_name, type);
